refactor database.js into separate modules (#108).
